requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
html2text>=2020.1.16
python-slugify>=8.0.0
